third_party/WebKit/LayoutTests/http/tests/notifications/serviceworkerregistration-service-worker-get-filter.html

<!doctype html>
<html>
  <head>
    <title>Notifications: ServiceWorkerRegistration.getNotifications() within a Service Worker with a filter.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the getNotifications() function when used in a Service Worker
      // return an array of the notifications which were previously displayed using
      // the same Service Worker registration id.
      async_test(function(test) {
          var scope = 'resources/scope/' + location.pathname,
              script = 'instrumentation-service-worker.js';

          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);

          var info = null;
          getActiveServiceWorkerWithMessagePort(test, script, scope).then(function(workerInfo) {
              info = workerInfo;

              // (1) Display two notifications in the Document.
              return info.registration.showNotification('Hello, world!', {
                body: 'First notification',
                tag: 'banana',
              });
          }).then(function() {
              return info.registration.showNotification('Hello again, world!', {
                body: 'Second notification',
                tag: 'strawberry',
              });
          }).then(function() {
              // (2) Request the Service Worker to give us a filtered list of notifications.
              return sendCommand(info.port, {
                  command: 'get',
                  filter: {
                    tag: 'strawberry',
                  }
              });
          }).then(function(data) {
              // (3) Confirm that the Service Worker was able to get the one matching notification.
              assert_true(data.success);

              var notifications = data.notifications;

              assert_equals(notifications.length, 1);

              assert_equals(notifications[0].title, 'Hello again, world!');
              assert_equals(notifications[0].body, 'Second notification');

              test.done();
          }).catch(unreached_rejection(test));

      }, 'ServiceWorkerRegistration.getNotifications() returns the opened notifications within a Service Worker with a filter.');
    </script>
  </body>
</html>
